+++ projects/rocdbgapi/src/agent.cpp
@@ -60,6 +60,8 @@ agent_t::agent_t (amd_dbgapi_agent_id_t agent_id, process_t &process,
           throw process_exited_exception_t (m_process.id ());
         else if (status == AMD_DBGAPI_STATUS_ERROR_MEMORY_ACCESS)
           throw memory_access_error_t (agent_address_space (), address);
+        else if (status == AMD_DBGAPI_STATUS_ERROR_MEMORY_UNAVAILABLE)
+          throw memory_unavailable_error_t (agent_address_space (), address);
         else if (status != AMD_DBGAPI_STATUS_SUCCESS)
           fatal_error ("xfer_global_memory_partial failed (%s)",
                        to_cstring (status));